Level_3/requirements.txt
langchain
langchain-openai
langchain-mcp-adapters
langgraph
fastmcp
python-dotenv
openai  
tavily-python
deepagents
langchain-community 
langchain-core
wikipedia